Clean-up
- Move mixin to the correct place (registrations module)
- Move code that's not shared between both RHs out of the mixin
- Use explicit selector for management mode
